Extract Event into it's own file
